fix: rename EmptyState to index.tsx
