+++ src/components/WorkspaceView/index.tsx
@@ -10,7 +10,7 @@ import { useT } from '../../lib/i18n'
 import { cellStyle, gridContainerStyle, reconcileGridLayout } from '../../lib/gridLayout'
 import type { AgentType, GridLayout, Group, Project, Terminal, WorkspaceContainer } from '../../lib/types'
 import { AgentIcon } from '../icons/AgentIcons'
-import { EmptyState } from '../EmptyState/EmptyState'
+import { EmptyState } from '../EmptyState'
 import { PaneArea } from './PaneArea'
 import { ProjectContainer } from './ProjectContainer'
 import styles from './WorkspaceView.module.css'
